+++ app/(auth)/signup/(SignupSteps)/step1/index.tsx
@@ -1,6 +1,10 @@
 'use client'
 import React from 'react'
+import { format } from 'date-fns'
+import { CalendarIcon } from 'lucide-react'
 import * as z from 'zod'
+import { cn } from '@/lib/utils'
+import { Calendar } from '@/components/ui/calendar'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { Form, FormProvider, useForm } from 'react-hook-form'
 import {
@@ -12,16 +16,14 @@ import {
   FormLabel,
 } from '@/components/ui/form'
 import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group'
-import { Button } from '@/components/ui/Button/button'
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from '@/components/ui/popover'
+import { Button } from '@/components/ui/button'
 import Link from 'next/link'
 import { IoIosArrowBack } from 'react-icons/io'
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select'
 
 const formSchema = z.object({
   dob: z.date({ required_error: 'A date of birth is required.' }),
@@ -30,7 +32,7 @@ const formSchema = z.object({
   }),
 })
 
-const StepThree = () => {
+const StepOne = () => {
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: {},
@@ -46,11 +48,9 @@ const StepThree = () => {
         />
       </div>
       <div className="flex flex-col gap-3 mb-6">
-        <h1 className="text-2xl font-bold">
-          We look out for you in any way possible
-        </h1>
+        <h1 className="text-2xl font-bold">Let us get to know you better</h1>
         <p className="text-sm font-semibold">
-          Let us make this journey as convenient for you as we can...
+          We will use this information to personalize your experience
         </p>
       </div>
 
@@ -61,12 +61,54 @@ const StepThree = () => {
         >
           <FormField
             control={form.control}
+            name="dob"
+            render={({ field }) => (
+              <FormItem className="flex flex-col">
+                <FormLabel>What is your date of birth?</FormLabel>
+                <Popover>
+                  <PopoverTrigger asChild>
+                    <FormControl>
+                      <Button
+                        variant={'outline'}
+                        className={cn(
+                          'rounded-xl py-6 text-left font-normal',
+                          !field.value && 'text-muted-foreground',
+                        )}
+                      >
+                        {field.value ? (
+                          format(field.value, 'PPP')
+                        ) : (
+                          <span>Date of birth</span>
+                        )}
+                        <CalendarIcon className="ml-auto h-4 w-4 opacity-50" />
+                      </Button>
+                    </FormControl>
+                  </PopoverTrigger>
+                  <PopoverContent className="w-auto p-0" align="start">
+                    <Calendar
+                      mode="single"
+                      selected={field.value}
+                      onSelect={field.onChange}
+                      disabled={(date) =>
+                        date > new Date() || date < new Date('1900-01-01')
+                      }
+                      initialFocus
+                    />
+                  </PopoverContent>
+                </Popover>
+                <FormDescription>
+                  Your date of birth is used to calculate your age.
+                </FormDescription>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
             name="type"
             render={({ field }) => (
               <FormItem className="space-y-3">
-                <FormLabel>
-                  How do you rate yourself in terms of your current health?
-                </FormLabel>
+                <FormLabel>Please specify your gender</FormLabel>
                 <FormControl>
                   <RadioGroup
                     onValueChange={field.onChange}
@@ -77,25 +119,19 @@ const StepThree = () => {
                       <FormControl>
                         <RadioGroupItem value="all" />
                       </FormControl>
-                      <FormLabel className="font-normal">Excellent</FormLabel>
+                      <FormLabel className="font-normal">Male</FormLabel>
                     </FormItem>
                     <FormItem className="flex items-center space-x-3 space-y-0">
                       <FormControl>
                         <RadioGroupItem value="mentions" />
                       </FormControl>
-                      <FormLabel className="font-normal">Good</FormLabel>
+                      <FormLabel className="font-normal">Female</FormLabel>
                     </FormItem>
                     <FormItem className="flex items-center space-x-3 space-y-0">
                       <FormControl>
                         <RadioGroupItem value="none" />
                       </FormControl>
-                      <FormLabel className="font-normal">Fair</FormLabel>
-                    </FormItem>
-                    <FormItem className="flex items-center space-x-3 space-y-0">
-                      <FormControl>
-                        <RadioGroupItem value="none" />
-                      </FormControl>
-                      <FormLabel className="font-normal">Poor</FormLabel>
+                      <FormLabel className="font-normal">Other</FormLabel>
                     </FormItem>
                     <FormItem className="flex items-center space-x-3 space-y-0">
                       <FormControl>
@@ -116,7 +152,45 @@ const StepThree = () => {
             name="type"
             render={({ field }) => (
               <FormItem className="space-y-3">
-                <FormLabel>Do you have disability?</FormLabel>
+                <FormLabel>Please specify your relationship status</FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex space-y-1"
+                  >
+                    <FormItem className="flex items-center space-x-3 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="all" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Single</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-3 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="mentions" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Married</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-3 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="none" />
+                      </FormControl>
+                      <FormLabel className="font-normal">
+                        Prefer not to answer
+                      </FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="type"
+            render={({ field }) => (
+              <FormItem className="space-y-3">
+                <FormLabel>Do you have any kids?</FormLabel>
                 <FormControl>
                   <RadioGroup
                     onValueChange={field.onChange}
@@ -149,25 +223,6 @@ const StepThree = () => {
               </FormItem>
             )}
           />
-          <FormField
-            control={form.control}
-            name="dob"
-            render={({ field }) => (
-              <FormItem className="flex flex-col">
-                <FormLabel>Please select one from the options</FormLabel>
-                <Select>
-                  <SelectTrigger className="w-[180px]">
-                    <SelectValue placeholder="Select" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="light">Light</SelectItem>
-                    <SelectItem value="dark">Dark</SelectItem>
-                    <SelectItem value="system">System</SelectItem>
-                  </SelectContent>
-                </Select>
-              </FormItem>
-            )}
-          />
           <div className=" flex justify-end gap-3">
             <Button
               type="submit"
@@ -179,7 +234,7 @@ const StepThree = () => {
               type="submit"
               className=" bg-[#217873] hover:bg-[#217873] px-8"
             >
-              <Link href="/">Next</Link>
+              <Link href="/signup/step3">Next</Link>
             </Button>
           </div>
         </form>
@@ -188,4 +243,4 @@ const StepThree = () => {
   )
 }
 
-export default StepThree
+export default StepOne
