+++ app/(auth)/signup/(SignupSteps)/step1/page.tsx
@@ -1,0 +1,7 @@
+import StepOne from '.'
+
+const page = () => {
+  return <StepOne />
+}
+
+export default page
