--- app/(auth)/signup/(SignupSteps)/step3/page.tsx
@@ -1,7 +1,0 @@
-import StepThree from '.'
-
-const page = () => {
-  return <StepThree />
-}
-
-export default page
